Rename folder to “incompatible-extension-notice”
