+++ assets/js/editor-components/incompatible-extension-notice/editor.scss
