+++ assets/js/editor-components/incompatible-extension-notice/index.tsx
